+++ src/main/java/com/intern/hub/pm/dto/task/TaskResponse.java
@@ -3,6 +3,7 @@ package com.intern.hub.pm.dto.task;
 import com.intern.hub.pm.dto.document.DocumentResponse;
 import com.intern.hub.pm.model.constant.StatusWork;
 
+import java.math.BigInteger;
 import java.time.LocalDateTime;
 import java.util.List;
 
@@ -14,7 +15,7 @@ public record TaskResponse(
         String description,
         String note,
         StatusWork status,
-        Long rewardToken,
+        BigInteger rewardToken,
         String creatorId,
         String assigneeId,
         String creatorName,
