+++ src/main/java/com/intern/hub/pm/dto/task/TaskUpsertRequest.java
@@ -6,12 +6,13 @@ import jakarta.validation.constraints.Min;
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
 
+import java.math.BigInteger;
 import java.time.LocalDateTime;
 
 public record TaskUpsertRequest(
         @NotBlank String name,
         String description,
-        @NotNull @Min(0) Long rewardToken,
+        @NotNull @Min(0) BigInteger rewardToken,
         @NotNull @JsonSerialize(using = ToStringSerializer.class) Long assigneeId,
         @NotNull LocalDateTime startDate,
         @NotNull LocalDateTime endDate)
